==> 15.Meme-suite/1.get_fasta.ipynb <==
import pandas as pd

from Bio import SeqIO
from Bio.SeqRecord import SeqRecord
from Bio.Seq import Seq

data_LmSIDERs_pos = pd.read_csv("./data/positive_elements_LmSIDERs.csv", header=None, sep=",")
data_LmSIDERs_pos.columns = ["sseqid", "length", "sstart", "send", "sstrand", "sseq"]
print(data_LmSIDERs_pos.shape)
print(data_LmSIDERs_pos.dtypes)
data_LmSIDERs_pos.head()

def fasta_creator(csv_input, output_path):
    matrix = []
    for index, row in csv_input.iterrows():
        rec = SeqRecord(Seq(row["sseq"]), 
                        id = f"Seq_{index}_{row['sseqid']}",
                        description = "Leishmania infantum"
                        )
        matrix.append(rec)
    SeqIO.write(matrix, output_path, "fasta")

fasta_creator(data_LmSIDERs_pos, "./data/positive_elements_LmSIDERs.fasta")
