==> 15.Meme-suite/2.join_strands.ipynb <==
# Needed modules
import pandas as pd
import os
import subprocess

# Load the consensus data output
consensus_data = pd.read_csv("./data_software_output/Last_Data.csv", sep=",", header=0)
print(consensus_data.shape)
print(consensus_data.dtypes)
consensus_data.head()

# Check if there are rows with sstart > send
sum(consensus_data["sstart"] > consensus_data["send"])

# Make it so that sstart < send
consensus_data.loc[consensus_data["sstrand"] == "minus", ["sstart", "send"]] = consensus_data.loc[consensus_data["sstrand"] == "minus", ["send", "sstart"]].values

# Check it it worked
sum(consensus_data["sstart"] > consensus_data["send"])

# Take only what's needed
consensus_data = consensus_data[["sseqid", "sstart", "send"]].copy()

# Path for bedops files creawtion
path_consensus = "./data_software_output/consensus_seqs.bed"
consensus_data.to_csv(path_consensus, sep="\t", header=False, index=False)  # save data for BEDOPS file

# Bedops call
cmd = f"bedops --merge {path_consensus}"
consensus_data = subprocess.check_output(cmd, shell=True, universal_newlines=True)
consensus_data = pd.DataFrame([x.split("\t") for x in consensus_data.split("\n") if x],
                              columns=["sseqid", "sstart", "send"])
consensus_data[["sstart", "send"]] = consensus_data[["sstart", "send"]].apply(pd.to_numeric)
print(f"There are {consensus_data.shape[0]} consensus sequences after BEDOPS merge")

# Now let's get the sequences
def get_data_sequence(data, strand, genome_fasta):
    sequences = []
    for _, row in data.iterrows():
        sseqid = row["sseqid"]
        start = row["sstart"]
        end = row["send"]
        cmd = f"blastdbcmd -db {genome_fasta} -entry {sseqid} -range {start}-{end} -strand {strand} -outfmt %s"

        sequence = subprocess.check_output(cmd, shell=True, universal_newlines=True).replace('\n', '')

        sequences.append({
            "sseqid": sseqid,
            "sstart": start,
            "send": end,
            "sstrand": strand,
            "sseq": sequence
        })

    sequences_df = pd.DataFrame(sequences)
    return sequences_df

# Prepare dict creation
def blastn_dic(path_input, path_output):
    # "parse_seqids" is used to keep the sequence ID in the output.
    cmd = f"makeblastdb -in {path_input} -dbtype nucl -parse_seqids -out {path_output}"
    subprocess.run(cmd, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)

# Prepare dictionary blastn
os.makedirs("./data/L_infantum", exist_ok=True)
blastn_dic(path_input="../0.Data/genome/L_infantum/TriTrypDB-67_LinfantumJPCM5_Genome.fasta",
           path_output="./data/L_infantum/TriTrypDB-67_LinfantumJPCM5_Genome.fasta")

consensus_data_seqs = get_data_sequence(data=consensus_data,
                                        strand="plus",
                                        genome_fasta="./data/L_infantum/TriTrypDB-67_LinfantumJPCM5_Genome.fasta"
                                        )
print(f"There are {consensus_data_seqs.shape[0]} consensus sequences after getting the sequences")

# explore data
print(consensus_data_seqs.shape)
print(consensus_data_seqs.dtypes)
consensus_data_seqs.head()

# Insert length column
length_column = consensus_data_seqs["sseq"].apply(len)
consensus_data_seqs.insert(1, "length", length_column)
print(consensus_data_seqs.dtypes)
consensus_data_seqs.head()

consensus_data_seqs.to_csv("./data_software_output/consensus_data_seqs.csv", sep=",", index=False, header=True)
