==> 15.Meme-suite/3.sensibility.ipynb <==
import pandas as pd
import subprocess
import os

# Consensus vs TP data
In this first case we are not going to use the filtering methods

# Load TP data
TP_data = pd.read_csv("./data/TP_SIDERs_coor.csv", sep=",", header=0)
print(TP_data.shape)
print(TP_data.dtypes)
TP_data.head()

# Check if "sstart" > "send"
sum(TP_data["sstart"] > TP_data["send"])

# Let's correct it
TP_data.loc[TP_data["sstrand"] == "minus", ["sstart", "send"]] = TP_data.loc[TP_data["sstrand"] == "minus", ["send", "sstart"]].values

# Check if "sstart" > "send", there should be 0
sum(TP_data["sstart"] > TP_data["send"])

# Load consensus data
consensus_data = pd.read_csv("./data_software_output/consensus_data_seqs.csv", sep=",", header=0)
print(consensus_data.shape)
print(consensus_data.dtypes)
consensus_data.head()

# Check in consensus data if there are rows with "sstart" > "send"
sum(consensus_data["sstart"] > consensus_data["send"])

# Group data by chromosomes
TP_data_grouped = TP_data.groupby("sseqid")
consensus_data_grouped = consensus_data.groupby("sseqid")

path_1 = "./compare/consensus_vs_TP"
compare_dict_1 = {}
Total = 0
for (name1, group1), (name2, group2) in zip(TP_data_grouped, consensus_data_grouped):
    # group1 should be the TP data
    # ------------------------------------------------------------------------------
    path_chr = os.path.join(path_1, name1)
    os.makedirs(path_chr, exist_ok=True)
    # ------------------------------------------------------------------------------
    group1_total = group1[["sseqid", "sstart", "send"]].copy()
    group2_total = group2[["sseqid", "sstart", "send"]].copy()
    # ------------------------------------------------------------------------------
    group1_total.sort_values(by=["sstart", "send"], inplace=True)
    group2_total.sort_values(by=["sstart", "send"], inplace=True)
    # ------------------------------------------------------------------------------
    path_group1_total = os.path.join(path_chr, "group1_total.bed")
    path_group2_total = os.path.join(path_chr, "group2_total.bed")
    # ------------------------------------------------------------------------------
    group1_total.to_csv(path_group1_total, sep="\t", header=False, index=False)  # tabular sep because of bed format
    group2_total.to_csv(path_group2_total, sep="\t", header=False, index=False)  # tabular sep because of bed format

    # ------------------------------------------------------------------------------
    condition = 1
    cmd = f"bedops --element-of {condition} {path_group1_total} {path_group2_total}"
    result = subprocess.check_output(cmd, shell=True, universal_newlines=True)
    result_df = pd.DataFrame([x.split("\t") for x in result.split("\n") if x], 
                                columns=["sseqid", "sstart", "send"])
    result_df[['sstart', 'send']] = result_df[['sstart', 'send']].apply(pd.to_numeric)
    compare_dict_1[name1] = [f"{result_df.shape[0]}/{group1_total.shape[0]}", f"{result_df.shape[0]/group1_total.shape[0]*100:.2f}"]
    Total += result_df.shape[0]
print(f"""
There are {Total} TP sequences of {TP_data.shape[0]} in consensus data:
    - That's {Total/TP_data.shape[0]*100:.2f}% of the TP data
    - {TP_data.shape[0]-Total} TP sequences are not in consensus data, which is {100-Total/TP_data.shape[0]*100:.2f}%
"""
)
compare_dict_1

# Consensus vs 32 chr data

From here we need the filtering methods:

1. Use `3.1.SIDERs_filter.py`file
2. Filter the data from the consensus software.
3. Keep with this script

# Let's loead the data from the 32 chr
chr32_positive_data = pd.read_csv("../14.From_all_to_real_SIDERs/data//4.SIDERs_filter_32chr/positives_testing_elements.csv", sep=",", header=None)
chr32_positive_data.columns = ["sseqid", "length", "sstart", "send", "sstrand", "sseq"]
print(chr32_positive_data.shape)
print(chr32_positive_data.dtypes)
chr32_positive_data.head()

# Load the positive after the filter data from consensus sequences
consensus_positive_data = pd.read_csv("./data_software_output/positives_testing_elements.csv", sep=",", header=0)
print(consensus_positive_data.shape)
print(consensus_positive_data.dtypes)
consensus_positive_data.head()

# Let's group it
chr32_positive_data_grouped = chr32_positive_data.groupby("sseqid")
consensus_positive_data_grouped = consensus_positive_data.groupby("sseqid")

# Prepare path
path_2 = "./compare/consensus_vs_32chr"
os.makedirs(path_2, exist_ok=True)

# Let's define the function since it seems that we will use it more than once
def compare_sequences(df_group1, df_group2, df_group1_og_data, path):
    compare_dict = {}
    Total = 0
    for (name1, group1), (name2, group2) in zip(df_group1, df_group2):
        # group1 should be the TP data
        # ------------------------------------------------------------------------------
        path_chr = os.path.join(path_1, name1)
        os.makedirs(path_chr, exist_ok=True)
        # ------------------------------------------------------------------------------
        group1_total = group1[["sseqid", "sstart", "send"]].copy()
        group2_total = group2[["sseqid", "sstart", "send"]].copy()
        # ------------------------------------------------------------------------------
        group1_total.sort_values(by=["sstart", "send"], inplace=True)
        group2_total.sort_values(by=["sstart", "send"], inplace=True)
        # ------------------------------------------------------------------------------
        path_group1_total = os.path.join(path_chr, "group1_total.bed")
        path_group2_total = os.path.join(path_chr, "group2_total.bed")
        # ------------------------------------------------------------------------------
        group1_total.to_csv(path_group1_total, sep="\t", header=False, index=False)  # tabular sep because of bed format
        group2_total.to_csv(path_group2_total, sep="\t", header=False, index=False)  # tabular sep because of bed format

        # ------------------------------------------------------------------------------
        condition = 1
        cmd = f"bedops --element-of {condition} {path_group1_total} {path_group2_total}"
        result = subprocess.check_output(cmd, shell=True, universal_newlines=True)
        result_df = pd.DataFrame([x.split("\t") for x in result.split("\n") if x], 
                                    columns=["sseqid", "sstart", "send"])
        result_df[['sstart', 'send']] = result_df[['sstart', 'send']].apply(pd.to_numeric)
        compare_dict[name1] = [f"{result_df.shape[0]}/{group1_total.shape[0]}", f"{result_df.shape[0]/group1_total.shape[0]*100:.2f}"]
        Total += result_df.shape[0]
    print(f"""
    There are {Total} first DF sequences of {df_group1_og_data.shape[0]} in the second DF:
        - That's {Total/df_group1_og_data.shape[0]*100:.2f}% of the TP data
        - {df_group1_og_data.shape[0]-Total} TP sequences are not in consensus data, which is {100-Total/df_group1_og_data.shape[0]*100:.2f}%
    """
    )
    return compare_dict

consensus_vs_32chr = compare_sequences(df_group1=chr32_positive_data_grouped,  # compare if this data is inside.
                                       df_group2=consensus_positive_data_grouped,  # ... this data
                                       df_group1_og_data=chr32_positive_data,  # ... and this is the original data
                                       path=path_2)
consensus_vs_32chr

# Consensus + LmSIDERs vs TP data

# Let's load the dat from the LmSIDERs
LmSIDERs_data = pd.read_csv("../14.From_all_to_real_SIDERs/data/2.SIDERs_filter/1.Merged_data.csv", sep=",", header=0)
print(LmSIDERs_data.shape)
print(LmSIDERs_data.dtypes)
LmSIDERs_data.head()

# Let's check if there are rows with "sstart" > "send"
sum(LmSIDERs_data["sstart"] > LmSIDERs_data["send"])

# Let's make the path
path_3 = "./compare/consensus+LmSIDERs_vs_TP"
os.makedirs(path_3, exist_ok=True)

print(f"""
consensus data is {consensus_data.shape[0]} long
LmSIDERs data is {LmSIDERs_data.shape[0]} long
""")

# prepare paths for bedos merging
consensus_bed_path = os.path.join(path_3, "consensus.bed")
LmSIDERs_bed_path = os.path.join(path_3, "LmSIDERs.bed")

# prepare bedops file for the merging
consensus_data[["sseqid", "sstart", "send"]].to_csv(consensus_bed_path, sep="\t", header=False, index=False)
LmSIDERs_data[["sseqid", "sstart", "send"]].to_csv(LmSIDERs_bed_path, sep="\t", header=False, index=False)

# Let's BEDOPS merge both data
cmd =f"bedops --merge {consensus_bed_path} {LmSIDERs_bed_path}"
consensus_LMSIDERs_merged = subprocess.check_output(cmd, shell=True, universal_newlines=True)
consensus_LMSIDERs_merged_df = pd.DataFrame([x.split("\t") for x in consensus_LMSIDERs_merged.split("\n") if x], 
                                            columns=["sseqid", "sstart", "send"])
consensus_LMSIDERs_merged_df[['sstart', 'send']] = consensus_LMSIDERs_merged_df[['sstart', 'send']].apply(pd.to_numeric)
print(consensus_LMSIDERs_merged_df.shape)
print(consensus_LMSIDERs_merged_df.dtypes)
print(f"consensus_LmSIDERs_merged_df is {consensus_LMSIDERs_merged_df.shape[0]} long")
consensus_LMSIDERs_merged_df.head()


# Let's get the sequences
# Now let's get the sequences
def get_data_sequence(data, strand, genome_fasta):
    sequences = []
    for _, row in data.iterrows():
        sseqid = row["sseqid"]
        start = row["sstart"]
        end = row["send"]
        cmd = f"blastdbcmd -db {genome_fasta} -entry {sseqid} -range {start}-{end} -strand {strand} -outfmt %s"

        sequence = subprocess.check_output(cmd, shell=True, universal_newlines=True).replace('\n', '')

        sequences.append({
            "sseqid": sseqid,
            "sstart": start,
            "send": end,
            "sstrand": strand,
            "sseq": sequence
        })

    sequences_df = pd.DataFrame(sequences)
    return sequences_df

# Let's get the sequences
consensus_LMSIDERs_seqs = get_data_sequence(data=consensus_LMSIDERs_merged_df,
                                            strand="plus",
                                            genome_fasta="./data/L_infantum/TriTrypDB-67_LinfantumJPCM5_Genome.fasta")
print(consensus_LMSIDERs_seqs.shape)
print(consensus_LMSIDERs_seqs.dtypes)
consensus_LMSIDERs_seqs.head()

# Save the data
path_consensus_LMSIDERs_seqs = os.path.join(path_3, "consensus_LmSIDERs_seqs.csv")
consensus_LMSIDERs_seqs.to_csv(path_consensus_LMSIDERs_seqs, sep=",", header=True, index=False)

# Let's group this last data
consensus_LMSIDERs_seqs_grouped = consensus_LMSIDERs_seqs.groupby("sseqid")

consensus_LMSIDERs_vs_TP = compare_sequences(df_group1=TP_data_grouped,  # compare if this data is inside.
                                             df_group2=consensus_LMSIDERs_seqs_grouped,  # ... this data
                                             df_group1_og_data=TP_data,  # ... and this is the original data
                                             path=path_3)
consensus_LMSIDERs_vs_TP

# Consensus + LmSIDERS vs 32 chr data

# Prepare paths
path_4 = "./compare/consensus+LmSIDERs_vs_32chr"
os.makedirs(path_4, exist_ok=True)

From here we need the filtering methods:

1. Use `3.1.SIDERs_filter.py`file
2. Filter the data from consensus + LmSIDERs.
3. Keep with this script

# Load data after the filter
consensus_LmSIDERS_positive_data = pd.read_csv("./compare/consensus+LmSIDERs_vs_TP/positives_testing_elements.csv", sep=",", header=0)
print(consensus_LmSIDERS_positive_data.shape)
print(consensus_LmSIDERS_positive_data.dtypes)
consensus_LmSIDERS_positive_data.head()

# Group this data
consensus_LmSIDERS_positive_data_grouped = consensus_LmSIDERS_positive_data.groupby("sseqid")

consensus_LMSIDERs_vs_32chr = compare_sequences(df_group1=chr32_positive_data_grouped,  # compare if this data is inside.
                                                df_group2=consensus_LmSIDERS_positive_data_grouped,  # ... this data
                                                df_group1_og_data=chr32_positive_data,  # ... and this is the original data
                                                path=path_4)
consensus_LMSIDERs_vs_32chr

# LmSIDERs check

# Merged data without the last filtering
print(LmSIDERs_data.shape)
print(LmSIDERs_data.dtypes)
print(sum(LmSIDERs_data["sstart"] > LmSIDERs_data["send"]))
LmSIDERs_data.head()

# Merged data after the last filtering
LmSIDERs_positive_data = pd.read_csv("../14.From_all_to_real_SIDERs/data/2.SIDERs_filter/positives_testing_elements.csv", sep=",", header=None)
LmSIDERs_positive_data.columns = ["sseqid", "length", "sstart", "send", "sstrand", "sseq"]
print(LmSIDERs_positive_data.shape)
print(LmSIDERs_positive_data.dtypes)
print(sum(LmSIDERs_positive_data["sstart"] > LmSIDERs_positive_data["send"]))
LmSIDERs_positive_data.head()

# Set paths
path_5 = "./compare/LmSIDERs_vs"
os.makedirs(path_5, exist_ok=True)

# LmSIDERs not filtered vs TP:
LmSIDERs_vs_TP = compare_sequences(df_group1=TP_data_grouped,  # compare if this data is inside.
                                   df_group2=LmSIDERs_data.groupby("sseqid"),  # ... this data
                                   df_group1_og_data=TP_data,  # ... and this is the original data
                                   path=path_5)
LmSIDERs_vs_TP

# LmSIDERs filtered vs TP:
LmSIDERs_positive_vs_TP = compare_sequences(df_group1=TP_data_grouped,  # compare if this data is inside.
                                           df_group2=LmSIDERs_positive_data.groupby("sseqid"),  # ... this data
                                           df_group1_og_data=TP_data,  # ... and this is the original data
                                           path=path_5)
LmSIDERs_positive_vs_TP
